requirements.txt
pandas
nltk
streamlit
python-dotenv
langchain
langchain-core
langchain-groq